+++ multipaz-models/src/commonMain/kotlin/org/multipaz/models/openid/dcql/DcqlCredentialQueryException.kt
@@ -1,8 +1,0 @@
-package org.multipaz.models.openid.dcql
-
-/**
- * Thrown when a credential query cannot be satisfied.
- *
- * @param message error message with detail.
- */
-class DcqlCredentialQueryException(message: String): Exception(message)