+++ samples/testapp/src/commonMain/kotlin/org/multipaz/testapp/TestAppSettingsModel.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import kotlinx.io.bytestring.ByteString
 import org.multipaz.cbor.buildCborArray
-import org.multipaz.models.digitalcredentials.DigitalCredentials
+import org.multipaz.digitalcredentials.DigitalCredentials
 import kotlin.Boolean
 
 /**